requirements.txt
torch>=2.0
torchaudio
scikit-learn
numpy
matplotlib
seaborn
pyyaml
